Move circuit tests closer to what they are testing
